// src/app/model/order-model.ts
export class OrderModal {
    ordered_foragency: any;
    ordered_forfarmer: any;
    ordered_forexuser: any;
    ordered_by: any;
    on_the_behalf: any;
    farmer_id: any;
    pincode: any;
    state: any;
    district: any;
    price: any;
    category: any;
    products: any = [];
    units: any;
    quantity: any;
    total_price: any;
    total_order: any;
    expected_date: any;
    delivered_date: any;
    place: any;
    status: any;
    number: any;
    address: any;
}